requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pydantic_settings
sqlalchemy
asyncpg
alembic
passlib
python-jose
jinja2
python-multipart
fastapi-cache2[redis]
celery
flower
pillow
sqladmin
itsdangerous
pytest
pytest-asyncio
httpx
asgi_lifespan